+++ go.mod
@@ -1,5 +1,3 @@
-module go-sse-poc
-
 go 1.22.0
 
 require (
